Delete unused files and remove navigation pages
