--- src/pages/authentication/component/navpages/Inspiration.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import "..navpages/styles.css";
-
-export const Inspiration = () => {
-  return (
-    <div className="inspiration" id="inspiration">
-      <h2><span className="fancy">What inspires us</span></h2>
-      <p>
-        Our journey began as two employees of Banco Popular, selected for a unique opportunity at Holberton Coding School, fully sponsored to innovate in project-based learning. Inspired by Banco Popular's MiBanco app, we recognized its potential for enhancement. Our goal was clear: to reimagine MiBanco, making it as user friendly as possible but also injecting elements of gamification to incentivize user engagement. The result is a fusion of our banking expertise and coding skills, delivering a more intuitive and, rewarding experience for MiBanco users.
-      </p>
-    </div>
-  );
-}
